Squash pthreaded vos warnings
Trying to build vos in src/tvolser can produce a warning, at least on
amd64_linux26, since we cast to an int of the wrong size in SendFile and
ReceiveFile. Use intptr_t like we do everywhere else instead.

Reviewed-on: http://gerrit.openafs.org/1556
Tested-by: Andrew Deason <adeason@sinenomine.net>

+++ src/volser/vos.c
@@ -295,7 +295,7 @@ SendFile(usd_handle_t ufd, register struct rx_call *call, long blksize)
 	FD_SET((intptr_t)(ufd->handle), &in);
 	/* don't timeout if read blocks */
 #if defined(AFS_PTHREAD_ENV)
-	select(((int)(ufd->handle)) + 1, &in, 0, 0, 0);
+	select(((intptr_t)(ufd->handle)) + 1, &in, 0, 0, 0);
 #else
 	IOMGR_Select(((intptr_t)(ufd->handle)) + 1, &in, 0, 0, 0);
 #endif
@@ -410,7 +410,7 @@ ReceiveFile(usd_handle_t ufd, struct rx_call *call, long blksize)
 	    FD_SET((intptr_t)(ufd->handle), &out);
 	    /* don't timeout if write blocks */
 #if defined(AFS_PTHREAD_ENV)
-	    select(((int)(ufd->handle)) + 1, &out, 0, 0, 0);
+	    select(((intptr_t)(ufd->handle)) + 1, &out, 0, 0, 0);
 #else
 	    IOMGR_Select(((intptr_t)(ufd->handle)) + 1, 0, &out, 0, 0);
 #endif
